Reorganize docs and examples
