+++ .gitignore
@@ -3,3 +3,4 @@
 /*/out
 /html
 /latex
+**/build/*
