+++ examples/When/CMakeLists.txt
@@ -5,4 +5,3 @@ project(WhenCpp)
 set(CMAKE_CXX_STANDARD 17)
 
 add_executable(when main.cpp)
-add_executable(when2 main2.cpp)
